work pending

+++ app/models/CodedOutDetail.scala
@@ -21,7 +21,7 @@ import play.api.libs.json.{Json, OFormat}
 import java.time.LocalDate
 
 case class CodedOutDetail(
-    totalAmount: Double,
+    totalAmount: BigDecimal,
     effectiveStartDate: LocalDate,
     effectiveEndDate: LocalDate
 )
